add logic to upload image from file
This commit has the code required to successfully upload an image to one of the repos
registered with glint. Currently there is only support for a direct upload but logic
for uploading from a url should follow shortly.

This code still needs a lot of error handling as a failure causes the image collection
thread to be locked out of 'dormant' mode. This can be accomplished by a simple try
except block that decrements the number of transactions. More error handling should
also be added regarding user feedback when something has gone wrong with the upload.

The uploaded image should also be registered with the image cache such that it can
be removed after all of the uploads have finished. Right now it may cause errors to
delete the image after upload since we may be uploading to multiple repos.

The image cache needs a lot of fleshing out to be used for transfers and a routine
to clean up old files still needs to be developped and put into place.

+++ glintwebui/glint_api.py
@@ -71,16 +71,26 @@ class repo_connector(object):
 		return image.id
 
 	# Upload an image to repo, returns True if successful or False if not
-	def upload_image(self, image_id, image_name, scratch_dir):
-		glance = glanceclient.Client('2', session=self.sess)
-		images = glance.images.list()
-		file_path = scratch_dir + image_name
-		glance.images.upload(image_id, open(file_path, 'rb'))
-		# Delete the file when we are done with it
-		logger.info("Upload complete, deleting temp file")
-		os.remove(file_path)
-		os.rmdir(scratch_dir)
-		return True
+	# if there is no image_id it is a direct upload and no placeholder exists
+	def upload_image(self, image_id, image_name, scratch_dir, disk_format=None, container_format=None):
+		if image_id is not None:
+			#this is the 2nd part of a transfer not a direct upload
+			glance = glanceclient.Client('2', session=self.sess)
+			images = glance.images.list()
+			file_path = scratch_dir + image_name
+			glance.images.upload(image_id, open(file_path, 'rb'))
+			# Delete the file when we are done with it
+			logger.info("Upload complete, deleting temp file")
+			os.remove(file_path)
+			os.rmdir(scratch_dir)
+			return True
+		else:
+			#this is a straight upload not part of a transfer
+			glance = glanceclient.Client('2', session=self.sess)
+			image = glance.images.create(name=image_name, disk_format=disk_format, container_format=container_format)
+			glance.images.upload(image.id, open(scratch_dir, 'rb'))
+			logger.info("Upload complete")
+			return True
 
 	# Download an image from the repo, returns True if successful or False if not
 	def download_image(self, image_name, image_id, scratch_dir):
